week5/treeofthoughtassignment/requirements.txt
openai
python_dotenv
langchain
langchain_openai
